FollowMatchersVisitor: don't attempt to collect following matchers...
Which means, in match() implementations, we won't need to context.setIntTag() or
anything.

A SequenceMatcher will basically try and collect all of its matchers until one
cannot match an empty sequence; it can work for it, but it cannot work for a
JoinMatcher, nor can it, for that matter, for a {Zero,One}OrMoreMatcher (they
always return false here).

The difference is those two collect their submatchers; JoinMatcher won't. This
is because JoinMatcher is potentially unbounded!