SystemOutCharSink: implement a custom Writer for System.out
CharSink seems to .close() its Writer after each .write*() operation which seems
to be somewhat overkill. I take it this is probably a bug but who knows. The
fact is that CharSink does not implement Closeable so maybe that is why they
chose to do it this way...

Anyway, this solves issue #2.