Reference: add .getNonnull()
Reference allows for null, which sucks. This will certainly change for 1.1 (with
the current Reference renamed to NullableReference).

However, since JSR 305 is used, .get() is annotated as @Nullable and this causes
concerns for people who emit warnings at compile time when this annotation is
present.

Introduce a .getNonnull() method which is annotated as @Nonnull; if the
underlying value is null, throw an IllegalStateException.

Fixes issue #9.