Improve the performance of case insensitive tries
For this, stop using a CharBuffer to walk through characters; instead, use a
CharSequence.

This means that the TrieBuilder now takes a boolean as an argument telling
whether the trie to be generated is case insensitive (true) or not (false).

As a result we need not convert the matched input to uppercase when matching,
resulting in significant performance improvements.